% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{bind_se_with_na}
\alias{bind_se_with_na}
\title{Bind SE objects - handle NAs for non-overlapping samples}
\usage{
bind_se_with_na(platforms, sample_id)
}
\arguments{
\item{platforms}{List of individual platform SEs.}

\item{sample_id}{Column name of the sample identifier.}
}
\value{
A SummarizedExperiment object with all platforms data
}
\description{
Binds the SummarizedExperiments of individual platforms into a single SummarizedExperiment where
all samples are kept and NAs are (purposefully) produced where a sample doesn tn have an
analyte measurement.
Note that the sample names must be in the same format and under the same column name for all
platforms!
}
