% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_R_from_SE}
\alias{initialize_R_from_SE}
\title{Initialize autofocus data struct from SummarizedExperiment}
\usage{
initialize_R_from_SE(
  SE,
  confounders,
  phenotype,
  cores = 4,
  use_wgcna = F,
  adj_method = c("bonferroni", "fdr"),
  corr_method = c("pearson", "spearman"),
  calculate_mgms = T
)
}
\arguments{
\item{SE}{SummarizedExperiment where assya is a matrix with samples as rows and biomolecules as columns, colData is a matrix of annotation data of samples, and rowData is a matrix of annotation data of biomolecules.}

\item{confounders}{Vector of names of potential confounders (names must be in colData of SE).}

\item{phenotype}{Name of column containing phenotype of interest, must be in colData of SE.}

\item{cores}{Number of cores to run on. Default: 4.}

\item{use_wgcna}{Whether or not to use WGNA for hiearchical clustering. Default: F.}

\item{adj_method}{Which p-value adjustment method to use. One of c("bonferroni", "fdr").}

\item{corr_method}{Which correlation method to use. One of c("pearson", "spearman")}

\item{calculate_mgms}{Boolean, should mgms be calculated (set to F for time saving purposes)}
}
\value{
R struct with the data matrix, correlation matrix,
sample and molecular annotations, hierarchical structure, cluster membership
node order within the hierarchical structure and a color vector for the platforms
}
\description{
Takes as input a data matrix, annotations on the samples and annotations on the molecules
and creates a list structure, R with the hierarchical clustering structure and cluster informations.
}
\author{
AS
}
