% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_se.R
\name{merge_se}
\alias{merge_se}
\title{Merge List of SE objects}
\usage{
merge_se(D_list, samp_id, feat_id, phenos, confounders)
}
\arguments{
\item{D_list}{List of SE objects.}

\item{samp_id}{Name of colData column containing sample identifiers. Must be the same for each
SE object.}

\item{feat_id}{Name of rowData column containing feature identifiers. Must be the same for each
SE object.}

\item{phenos}{A vector containing the phenotypes of interest. These must be colData columns and
they must be present in each SE object. This function will crash if these are not present.}

\item{confounders}{A vector containing confounders. These must be colData columns and they must
be present in each SE object. This function will crash if these are not present.}
}
\value{
Merges list of SE objects into a single large SE object.
}
\description{
Merges a list of SE objects provided/ Combines assay, rowData, and colData data frames. Each
SE object must satisfy the following requirements:
\itemize{
   \item{the name of the colData column containing sample identifiers must be the same across all SE objects}
   \item{the name of the rowData column containing feature identifiers must be the same across all SE objects}
   \item{samples can overlap, but all features must be unique}
   \item{the column names in \code{phenos} and \code{confounders} must be present in the colData data frame of each SE object}
}
}
\examples{
\dontrun{D_all <-
  # merge list of SE objects
  merge_se(D_list = list(D1, D2),
  samp_id="Sample_Id",
  feat_id="names",
  phenos="DIAB",
  confounders=c("AGE", "SEX", "BMI")) \%>\%
  ...}

}
\author{
KC
}
