% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_cluster_table.R
\name{output_cluster_table}
\alias{output_cluster_table}
\title{Output AutoFocus clusters at a specified threshold in table format}
\usage{
output_cluster_table(
  R,
  threshold,
  phenotype = "both",
  name,
  annotations_to_show
)
}
\arguments{
\item{R}{R_struct object output by initialize_R}

\item{threshold}{Enrichment threshold at which to calculate cluster enrichment}

\item{phenotype}{Which phenotype associated clusters to look at. Default "both" to see all phenotypes' clusters}

\item{name}{Column name in R$annos specifying the node name}

\item{annotations_to_show}{Vector of colnames of R$annos to return with nodes}
}
\value{
Dataframe with the nodes in clusters, their p-value association with phenotypes of interest, cluster, and annotations_to_show
}
\description{
Outputs a dataframe containing the information on nodes found in AutoFocus clusters of size greater than 1
}
\author{
AS
}
