% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_autofocus.R
\name{run_autofocus}
\alias{run_autofocus}
\title{Launch shiny app}
\usage{
run_autofocus(R, anno_list = c(""))
}
\arguments{
\item{R}{R struct}

\item{anno_list}{Vector of annotations to extract from R struct}
}
\value{
None (launches app)
}
\description{
Main autofocus function to launch shiny app.
}
