% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_analysis.R
\name{threshold_analysis}
\alias{threshold_analysis}
\title{Perform threshold analysis}
\usage{
threshold_analysis(R, phenotype, phenotype_name = NA)
}
\arguments{
\item{R}{R_struct object output by initialize_R}

\item{phenotype}{The phenotype for which to perform the threshold analysis (clusters are phenotype dependent)}

\item{phenotype_name}{Formal name of phenotype for plotting title}
}
\value{
None
}
\description{
Outputs a plot across thresholds from 0.01 to 0.99 at 0.01 increments with the following information:
\itemize{
   \item{Cluster Height Range: The height range of returned clusters at that threshold, normalized to the largest range across all thresholds}
   \item{Cluster Size Range: The size range of returned clusters at that threshold, normalized to the largest range across all thresholds}
   \item{Cluster Count: The number of returned clusters at that threshold, normalized to the highest number of clusters across all thresholds}
   \item{Non-Significant Node Proportion: The proportion of nodes in returned clusters that are not significantly associated with the phenotype of interest, normalized to the largest proportion across all thresholds}
   \item{Singleton Node Count: The number of nodes significantly associated to phenotype that are not found in a cluster, normalized to the largest number of singletons across all thresholds}
   \item{Noise Minimizing Score: A metric that combines the non-significant node proportion and singleton node count, to show at which threshold they are minimized}
}
}
\author{
AS
}
