% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_R}
\alias{initialize_R}
\title{Initialize autofocus data struct}
\usage{
initialize_R(
  data.matrix,
  sample.data,
  mol.data,
  confounders,
  phenotype,
  cores = 4,
  use_wgcna = F,
  adj_method = c("bonferroni", "fdr")
)
}
\arguments{
\item{data.matrix}{Matrix with samples as rows and biomolecules as columns.}

\item{sample.data}{Matrix of annotation data of samples.}

\item{mol.data}{Matrix of annotation data of biomolecules.}

\item{confounders}{Vector of names of potential confounders (names must be in sample.data).}

\item{phenotype}{Name of column containing phenotype of interest, must be in sample.data.}

\item{cores}{Number of cores to run on. Default: 4.}

\item{use_wgcna}{Whether or not to use WGNA for hiearchical clustering. Default: F.}

\item{adj_method}{Which p-value adjustment method to use. One of c("bonferroni", "fdr").}
}
\value{
R struct with the data matrix, correlation matrix,
sample and molecular annotations, hierarchical structure, cluster membership
node order within the hierarchical structure and a color vector for the platforms
}
\description{
Takes as input a data matrix, annotations on the samples and annotations on the molecules
and creates a list structure, R with the hierarchical clustering structure and cluster informations.
}
\author{
AS
}
